perf: don't include createdAt in L1 calldata
The createdAt timestamp isn't used in L1. Including it in the data that
is provided to the L1 contract (and also including it in the data that
is hashed and signed by the liquidity provider) is wasteful.

This adds an L1Ticket struct which includes only the required data,
which is the recipient and value of the transfer.